test: add shared context for admin logins
