+++ spec/support/shared_context.rb
@@ -1,7 +1,15 @@
 RSpec.shared_context 'logged in student', :shared_context => :metadata do
-  include StudentLoginHelper
+  include UserLoginHelper
 
   before(:each) do
     mock_current_student_login
   end 
+end
+
+RSpec.shared_context 'logged in admin', :shared_context => :metadata do
+  include UserLoginHelper
+
+  before(:each) do
+    mock_current_student_login_admin
+  end
 end
